playground: move playground views to new playground
