+++ playground/app/pages/circle.vue
@@ -1,16 +1,14 @@
 <script setup lang="ts">
-import { LMap, LTileLayer } from '@/maxel01/vue-leaflet'
-import { ref } from 'vue'
-
-const zoom = ref<number>(2)
+import { LCircle, LMap, LTileLayer } from '@maxel01/vue-leaflet'
 </script>
 
 <template>
-    <LMap v-model:zoom="zoom" :center="[47.41322, -1.219482]">
+    <LMap ref="map" :zoom="11" :center="[44.48865, 11.3317]">
         <LTileLayer
             url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
             layer-type="base"
             name="OpenStreetMap"
         />
+        <LCircle :lat-lng="[44.48865, 11.3317]" :radius="5000" color="green" />
     </LMap>
 </template>
