+++ playground/app/pages/control-zoom.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { LControlZoom, LMap, LTileLayer } from '@/maxel01/vue-leaflet'
+import { LControlZoom, LMap, LTileLayer } from '@maxel01/vue-leaflet'
 </script>
 
 <template>
