+++ playground/app/pages/control.vue
@@ -1,5 +1,5 @@
 <script setup>
-import { LMap, LTileLayer, LControl } from '@/maxel01/vue-leaflet'
+import { LMap, LTileLayer, LControl } from '@maxel01/vue-leaflet'
 
 const clickHandler = () => {
     alert('and mischievous')
