+++ playground/app/pages/feature-group.vue
@@ -1,14 +1,16 @@
 <script setup lang="ts">
-import { LCircle, LMap, LTileLayer } from '@/maxel01/vue-leaflet'
+import { LFeatureGroup, LMap, LMarker, LTileLayer } from '@maxel01/vue-leaflet'
 </script>
 
 <template>
-    <LMap ref="map" :zoom="11" :center="[44.48865, 11.3317]">
+    <LMap :zoom="2" :center="[47.41322, -1.219482]">
         <LTileLayer
             url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
             layer-type="base"
             name="OpenStreetMap"
         />
-        <LCircle :lat-lng="[44.48865, 11.3317]" :radius="5000" color="green" />
+        <LFeatureGroup>
+            <LMarker :lat-lng="[50, 50]" />
+        </LFeatureGroup>
     </LMap>
 </template>
