+++ playground/app/pages/polygon.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { LMap, LPolygon, LTileLayer } from '@/maxel01/vue-leaflet'
+import { LMap, LPolygon, LTileLayer } from '@maxel01/vue-leaflet'
 </script>
 
 <template>
