+++ playground/app/pages/polyline.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { LMap, LPolyline, LTileLayer } from '@/maxel01/vue-leaflet'
+import { LMap, LPolyline, LTileLayer } from '@maxel01/vue-leaflet'
 </script>
 
 <template>
